requirements.txt
rich>=13.0.0
aiosmtplib>=3.0.0
dnspython>=2.5.0
email-validator>=2.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
authlib>=1.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pyjwt>=2.8.0
asgiref>=3.7.2
gunicorn>=21.2.0
uvicorn>=0.27.0
